Added a generic DatabindingFragment to reduce binding boilerplate code Refactored all fragments to use it instead of manually overriding the inflater methods
