+++ app/src/main/java/de/thm/mobiletech/hideandguess/LoginFragment.kt
@@ -1,47 +1,16 @@
 package de.thm.mobiletech.hideandguess
 
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.databinding.DataBindingUtil
-import androidx.navigation.NavController
-import androidx.navigation.fragment.NavHostFragment
 import de.thm.mobiletech.hideandguess.databinding.FragmentLoginBinding
-import de.thm.mobiletech.hideandguess.databinding.FragmentMainMenuBinding
+import de.thm.mobiletech.hideandguess.util.DatabindingFragment
 
+class LoginFragment : DatabindingFragment<FragmentLoginBinding>(R.layout.fragment_login) {
 
-class LoginFragment : Fragment() {
-
-    private lateinit var binding: FragmentLoginBinding
-    private lateinit var navController: NavController
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        // Inflate layout using data binding
-        binding = DataBindingUtil.inflate(
-            inflater, R.layout.fragment_login, container, false
-        )
-        binding.lifecycleOwner = viewLifecycleOwner
-        binding.context = this // Set binding variable used in layout
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        // Find and assign navController
-        val navHostFragment = requireActivity().supportFragmentManager
-            .findFragmentById(R.id.navHost) as NavHostFragment
-        navController = navHostFragment.navController
+    override fun setBindingContext() {
+        binding.context = this
     }
 
     fun openRegisterFragment() {
         val action = LoginFragmentDirections.actionLoginFragmentToRegisterFragment()
         navController.navigate(action)
     }
-
 }
