+++ app/src/main/java/de/thm/mobiletech/hideandguess/MainActivity.kt
@@ -1,7 +1,7 @@
 package de.thm.mobiletech.hideandguess
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
