+++ app/src/main/java/de/thm/mobiletech/hideandguess/util/IDatabindingContext.kt
@@ -1,0 +1,18 @@
+package de.thm.mobiletech.hideandguess.util
+
+/**
+ * This interface enforces that all classes extending [DatabindingFragment]
+ * have to implement a method to set their data binding layout variable.
+ * Content of said method should be along the line of
+ *
+ * `binding.<contextVariable> = this`
+ * */
+sealed interface IDatabindingContext {
+    /**
+     * This method is called in [DatabindingFragment.onCreateView].
+     * It should set the data binding variable of the layout to the fragment instance.
+     *
+     * e.g. `binding.context = this`
+     * */
+    fun setBindingContext()
+}
